+++ packages/plugin-html/web/src/toolbar/HtmlEditPanel.jsx
@@ -1,8 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
-import { TextInput } from 'wix-rich-content-plugin-commons';
-import { RadioGroupHorizontal, InputWithLabel } from 'wix-rich-content-ui-components';
+import { RadioGroupHorizontal, InputWithLabel, TextInput } from 'wix-rich-content-ui-components';
 import { mergeStyles, isValidUrl, startsWithHttps, hasProtocol } from 'wix-rich-content-common';
 import { identity, trimStart } from 'lodash';
 import { SRC_TYPE_HTML, SRC_TYPE_URL } from '../defaults';